Move steps to separate files
